Complete fix 10

+++ ahrameev.vladislav/P3/matrix.cpp
@@ -31,44 +31,44 @@ void ahrameev::processSpiralDecrease(const int* src, size_t rows, size_t cols, s
     int val = 1;
     size_t count = 0;
 
-    size_t top = 0;
-    size_t bottom = rows - 1;
-    size_t left = 0;
-    size_t right = cols - 1;
+    int top = 0;
+    int bottom = static_cast<int>(rows) - 1;
+    int left = 0;
+    int right = static_cast<int>(cols) - 1;
 
     while (count < total)
     {
         // 1. Вверх по левому столбцу
-        for (size_t i = bottom; i >= top && count < total; --i)
+        for (int i = bottom; i >= top && count < total; --i)
         {
-            dst[i * cols + left] -= val++;
+            dst[i * static_cast<int>(cols) + left] -= val++;
             ++count;
         }
         ++left;
         if (count >= total) break;
 
         // 2. Вправо по верхней строке
-        for (size_t j = left; j <= right && count < total; ++j)
+        for (int j = left; j <= right && count < total; ++j)
         {
-            dst[top * cols + j] -= val++;
+            dst[top * static_cast<int>(cols) + j] -= val++;
             ++count;
         }
         ++top;
         if (count >= total) break;
 
         // 3. Вниз по правому столбцу
-        for (size_t i = top; i <= bottom && count < total; ++i)
+        for (int i = top; i <= bottom && count < total; ++i)
         {
-            dst[i * cols + right] -= val++;
+            dst[i * static_cast<int>(cols) + right] -= val++;
             ++count;
         }
         --right;
         if (count >= total) break;
 
         // 4. Влево по нижней строке
-        for (size_t j = right; j >= left && count < total; --j)
+        for (int j = right; j >= left && count < total; --j)
         {
-            dst[bottom * cols + j] -= val++;
+            dst[bottom * static_cast<int>(cols) + j] -= val++;
             ++count;
         }
         --bottom;
